refactor: P2 - BalanceService 인프라 의존성 제거, BalanceChangedEvent 도메인 이벤트 도입, 패키지 오타 수정
