+++ src/main/kotlin/com/hhplus/ecommerce/balance/domain/BalanceService.kt
@@ -1,25 +1,19 @@
 package com.hhplus.ecommerce.balance.domain
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.hhplus.ecommerce.common.anotation.RedisLock
-import com.hhplus.ecommerce.common.properties.BalanceKafkaProperties
 import com.hhplus.ecommerce.balance.domain.dto.BalanceQuery
 import com.hhplus.ecommerce.balance.domain.dto.BalanceResult
 import com.hhplus.ecommerce.balance.domain.dto.BalanceTransaction
-import com.hhplus.ecommerce.balance.domain.respository.IBalanceRepository
-import com.hhplus.ecommerce.balance.infrastructure.mongodb.BalanceHistoryDocument
-import com.hhplus.ecommerce.outboxevent.infrastructure.event.dto.OutboxEventInfo
+import com.hhplus.ecommerce.balance.domain.event.BalanceChangedEvent
+import com.hhplus.ecommerce.balance.domain.repository.IBalanceRepository
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.util.*
 
 @Service
 class BalanceService(
     private val balanceRepository: IBalanceRepository,
     private val applicationEventPublisher: ApplicationEventPublisher,
-    private val balanceKafkaProperties: BalanceKafkaProperties,
-    private val objectMapper: ObjectMapper,
 ) {
     fun validateBalanceToUse(item: BalanceTransaction) {
         val balanceEntity = balanceRepository.findByUserId(item.userId)
@@ -46,24 +40,16 @@ class BalanceService(
 
         balanceRepository.insertOrUpdate(balanceEntity)
 
-        val balanceHistoryDocument = BalanceHistoryDocument(
-            balanceId = balanceEntity.id,
-            amount = item.amount,
-            balance = balanceEntity.balance,
-            transactionType = "CHARGE"
+        applicationEventPublisher.publishEvent(
+            BalanceChangedEvent(
+                balanceId = balanceEntity.id,
+                amount = item.amount,
+                balance = balanceEntity.balance,
+                transactionType = "CHARGE"
+            )
         )
-
-        val outboxEvent = OutboxEventInfo(
-            id = UUID.randomUUID(),
-            groupId = balanceKafkaProperties.groupId,
-            topic = balanceKafkaProperties.topic,
-            payload = objectMapper.writeValueAsString(balanceHistoryDocument)
-        )
-
-        applicationEventPublisher.publishEvent(outboxEvent)
 
         return BalanceResult.from(balanceEntity)
-
     }
 
     @RedisLock(key = "'balance:' + #item.userId") // 잔액 정보에 대한 사용자 ID를 기반으로 Lock을 점유한다. 1:1 매핑이라 사용자 ID로 락을 잡도록 했다.
@@ -74,23 +60,14 @@ class BalanceService(
 
         balanceRepository.insertOrUpdate(balanceEntity)
 
-        println("잔액 정보 => ${balanceEntity.toString()}")
-
-        val balanceHistoryDocument = BalanceHistoryDocument(
-            balanceId = balanceEntity.id,
-            amount = item.amount,
-            balance = balanceEntity.balance,
-            transactionType = "USE"
+        applicationEventPublisher.publishEvent(
+            BalanceChangedEvent(
+                balanceId = balanceEntity.id,
+                amount = item.amount,
+                balance = balanceEntity.balance,
+                transactionType = "USE"
+            )
         )
-
-        val outboxEvent = OutboxEventInfo(
-            id = UUID.randomUUID(),
-            groupId = balanceKafkaProperties.groupId,
-            topic = balanceKafkaProperties.topic,
-            payload = objectMapper.writeValueAsString(balanceHistoryDocument)
-        )
-
-        applicationEventPublisher.publishEvent(outboxEvent)
 
         return BalanceResult.from(balanceEntity)
     }
