+++ src/main/kotlin/com/hhplus/ecommerce/balance/domain/event/BalanceChangedEvent.kt
@@ -1,0 +1,8 @@
+package com.hhplus.ecommerce.balance.domain.event
+
+data class BalanceChangedEvent(
+    val balanceId: Long,
+    val amount: Long,
+    val balance: Long,
+    val transactionType: String
+)
